+++ Assets/Architecture/Presentation/ApplicationEntryPoint.cs
@@ -1,6 +1,10 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
+using Cysharp.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using UnityEngine;
+using UnityEngine.AddressableAssets;
+using UnityEngine.SceneManagement;
 using VContainer.Unity;
 using ZLogger;
 
@@ -9,15 +13,23 @@ namespace Calc.Presentation
     internal sealed class ApplicationEntryPoint : IInitializable
     {
         private readonly ILogger<ApplicationEntryPoint> _logger;
+        private readonly AssetReference _mainScene;
+        private readonly MonoBehaviour _loader;
 
-        public ApplicationEntryPoint(ILogger<ApplicationEntryPoint> logger)
+        public ApplicationEntryPoint(ILogger<ApplicationEntryPoint> logger, AssetReference mainScene, MonoBehaviour loader)
         {
             _logger = logger;
+            _mainScene = mainScene;
+            _loader = loader;
         }
 
         void IInitializable.Initialize()
         {
             _logger.Info("Application is started!");
+
+            _mainScene.LoadSceneAsync(LoadSceneMode.Additive)
+                .ToUniTask(_loader)
+                .Forget(_logger.Exception);
         }
     }