+++ Assets/Architecture/Presentation/ViewsManagement/ErrorEvents.cs
@@ -1,0 +1,5 @@
+﻿namespace Calc.Presentation.ViewsManagement
+{
+    internal sealed record CalculationErrorOccured(string Message);
+    internal sealed record CalculationErrorProcessed { }
+}